[Writing Tools] Split up HTMLConverter into different files for each type of converter https://bugs.webkit.org/show_bug.cgi?id=296031 rdar://155935345
Reviewed by Aditya Keerthi.

The HTMLConverter.mm file has two implementations of HTML conversion, both of which are relatively complicated.
To improve readability and maintainability, split these implementations into their own files.

* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/editing/cocoa/EditingHTMLConverter.mm: Added.
(WebCore::preferredFilenameForElement):
(WebCore::fileWrapperForElement):
(WebCore::attributedStringWithAttachmentForFileWrapper):
(WebCore::attributedStringWithAttachmentForElement):
(WebCore::elementQualifiesForWritingToolsPreservation):
(WebCore::hasAncestorQualifyingForWritingToolsPreservation):
(WebCore::enclosingLinkElement):
(WebCore::updateAttributes):
(WebCore::editingAttributedStringInternal):
(WebCore::editingAttributedString):
(WebCore::editingAttributedStringReplacingNoBreakSpace):
* Source/WebCore/editing/cocoa/HTMLConverter.h:
* Source/WebCore/editing/cocoa/HTMLConverter.mm:
(WebCore::WebDefaultFont):
(WebCore::fileWrapperForURL):
(HTMLConverterCaches::isAncestorsOfStartToBeConverted const): Deleted.
(): Deleted.
(HTMLConverter::HTMLConverter): Deleted.
(HTMLConverter::convert): Deleted.
(WebDefaultFont): Deleted.
(_fontForNameAndSize): Deleted.
(defaultParagraphStyle): Deleted.
(HTMLConverterCaches::computedStylePropertyForElement): Deleted.
(HTMLConverterCaches::inlineStylePropertyForElement): Deleted.
(stringFromCSSValue): Deleted.
(HTMLConverterCaches::propertyValueForNode): Deleted.
(floatValueFromPrimitiveValue): Deleted.
(HTMLConverterCaches::floatPropertyValueForNode): Deleted.
(_shadowForShadowStyle): Deleted.
(HTMLConverterCaches::isBlockElement): Deleted.
(HTMLConverterCaches::elementHasOwnBackgroundColor): Deleted.
(HTMLConverter::_blockLevelElementForNode): Deleted.
(normalizedColor): Deleted.
(HTMLConverterCaches::colorPropertyValueForNode): Deleted.
(HTMLConverter::_colorForElement): Deleted.
(_font): Deleted.
(HTMLConverter::computedAttributesForElement): Deleted.
(HTMLConverter::attributesForElement): Deleted.
(HTMLConverter::aggregatedAttributesForAncestors): Deleted.
(HTMLConverter::aggregatedAttributesForElementAndItsAncestors): Deleted.
(HTMLConverter::_newParagraphForElement): Deleted.
(HTMLConverter::_newLineForElement): Deleted.
(HTMLConverter::_newTabForElement): Deleted.
(_WebMessageDocumentClass): Deleted.
(HTMLConverter::_addMultiRepresentationHEICAttachmentForImageElement): Deleted.
(HTMLConverter::_addAttachmentForElement): Deleted.
(HTMLConverter::_addQuoteForElement): Deleted.
(HTMLConverter::_addValue): Deleted.
(HTMLConverter::_fillInBlock): Deleted.
(read2DigitNumber): Deleted.
(_dateForString): Deleted.
(_colCompare): Deleted.
(HTMLConverter::_processMetaElementWithName): Deleted.
(HTMLConverter::_processHeadElement): Deleted.
(HTMLConverter::_enterBlockquote): Deleted.
(HTMLConverter::_exitBlockquote): Deleted.
(HTMLConverter::_enterElement): Deleted.
(HTMLConverter::_addLinkForElement): Deleted.
(HTMLConverter::_addTableForElement): Deleted.
(HTMLConverter::_addTableCellForElement): Deleted.
(HTMLConverter::_processElement): Deleted.
(HTMLConverter::_addMarkersToList): Deleted.
(HTMLConverter::_exitElement): Deleted.
(HTMLConverter::_processText): Deleted.
(HTMLConverter::_traverseNode): Deleted.
(HTMLConverter::_traverseFooterNode): Deleted.
(HTMLConverter::_adjustTrailingNewline): Deleted.
(HTMLConverterCaches::cacheAncestorsOfStartToBeConverted): Deleted.
(fileWrapperForURL): Deleted.
(preferredFilenameForElement): Deleted.
(fileWrapperForElement): Deleted.
(attributedStringWithAttachmentForFileWrapper): Deleted.
(attributedStringWithAttachmentForElement): Deleted.
(elementQualifiesForWritingToolsPreservation): Deleted.
(hasAncestorQualifyingForWritingToolsPreservation): Deleted.
(enclosingLinkElement): Deleted.
(updateAttributes): Deleted.
(WebCore::attributedString): Deleted.
(WebCore::editingAttributedStringInternal): Deleted.
(WebCore::editingAttributedString): Deleted.
(WebCore::editingAttributedStringReplacingNoBreakSpace): Deleted.
* Source/WebCore/editing/cocoa/NodeHTMLConverter.mm: Copied from Source/WebCore/editing/cocoa/HTMLConverter.mm.
(HTMLConverterCaches::isAncestorsOfStartToBeConverted const):
(HTMLConverter::HTMLConverter):
(HTMLConverter::convert):
(_fontForNameAndSize):
(defaultParagraphStyle):
(HTMLConverterCaches::computedStylePropertyForElement):
(HTMLConverterCaches::inlineStylePropertyForElement):
(stringFromCSSValue):
(HTMLConverterCaches::propertyValueForNode):
(floatValueFromPrimitiveValue):
(HTMLConverterCaches::floatPropertyValueForNode):
(_shadowForShadowStyle):
(HTMLConverterCaches::isBlockElement):
(HTMLConverterCaches::elementHasOwnBackgroundColor):
(HTMLConverter::_blockLevelElementForNode):
(normalizedColor):
(HTMLConverterCaches::colorPropertyValueForNode):
(HTMLConverter::_colorForElement):
(_font):
(HTMLConverter::computedAttributesForElement):
(HTMLConverter::attributesForElement):
(HTMLConverter::aggregatedAttributesForAncestors):
(HTMLConverter::aggregatedAttributesForElementAndItsAncestors):
(HTMLConverter::_newParagraphForElement):
(HTMLConverter::_newLineForElement):
(HTMLConverter::_newTabForElement):
(_WebMessageDocumentClass):
(HTMLConverter::_addMultiRepresentationHEICAttachmentForImageElement):
(HTMLConverter::_addAttachmentForElement):
(HTMLConverter::_addQuoteForElement):
(HTMLConverter::_addValue):
(HTMLConverter::_fillInBlock):
(read2DigitNumber):
(_dateForString):
(_colCompare):
(HTMLConverter::_processMetaElementWithName):
(HTMLConverter::_processHeadElement):
(HTMLConverter::_enterBlockquote):
(HTMLConverter::_exitBlockquote):
(HTMLConverter::_enterElement):
(HTMLConverter::_addLinkForElement):
(HTMLConverter::_addTableForElement):
(HTMLConverter::_addTableCellForElement):
(HTMLConverter::_processElement):
(HTMLConverter::_addMarkersToList):
(HTMLConverter::_exitElement):
(HTMLConverter::_processText):
(HTMLConverter::_traverseNode):
(HTMLConverter::_traverseFooterNode):
(HTMLConverter::_adjustTrailingNewline):
(HTMLConverterCaches::cacheAncestorsOfStartToBeConverted):
(WebCore::attributedString):

Canonical link: https://commits.webkit.org/297624@main
