+++ Source/WebCore/SaferCPPExpectations/UnretainedCallArgsCheckerExpectations
@@ -45,9 +45,10 @@ editing/SmartReplaceCF.cpp
 editing/cocoa/AttributedString.mm
 editing/cocoa/DataDetection.mm
 editing/cocoa/DictionaryLookup.mm
+editing/cocoa/EditingHTMLConverter.mm
 editing/cocoa/FontAttributeChangesCocoa.mm
 editing/cocoa/FontAttributesCocoa.mm
-editing/cocoa/HTMLConverter.mm
+editing/cocoa/NodeHTMLConverter.mm
 editing/cocoa/WebContentReaderCocoa.mm
 editing/mac/EditorMac.mm
 editing/mac/TextAlternativeWithRange.mm
