+++ Source/WebCore/SaferCPPExpectations/UnretainedLocalVarsCheckerExpectations
@@ -16,7 +16,8 @@ crypto/cocoa/SerializedCryptoKeyWrapMac.mm
 editing/cocoa/AlternativeTextUIController.mm
 editing/cocoa/AttributedString.mm
 editing/cocoa/DataDetection.mm
-editing/cocoa/HTMLConverter.mm
+editing/cocoa/EditingHTMLConverter.mm
+editing/cocoa/NodeHTMLConverter.mm
 loader/archive/cf/LegacyWebArchive.cpp
 loader/archive/cf/LegacyWebArchiveMac.mm
 loader/cocoa/BundleResourceLoader.mm
