+++ Source/WebCore/editing/cocoa/EditorCocoa.mm
@@ -46,7 +46,6 @@
 #import "FrameLoader.h"
 #import "FrameSelection.h"
 #import "HTMLAttachmentElement.h"
-#import "HTMLConverter.h"
 #import "HTMLImageElement.h"
 #import "HTMLPictureElement.h"
 #import "HTMLSourceElement.h"
@@ -56,6 +55,7 @@
 #import "LegacyNSPasteboardTypes.h"
 #import "LegacyWebArchive.h"
 #import "LocalFrame.h"
+#import "NodeHTMLConverter.h"
 #import "Page.h"
 #import "PagePasteboardContext.h"
 #import "Pasteboard.h"
