+++ Source/WebCore/editing/mac/EditorMac.mm
@@ -41,7 +41,6 @@
 #import "EditorClient.h"
 #import "FontAttributes.h"
 #import "FontShadow.h"
-#import "HTMLConverter.h"
 #import "HTMLElement.h"
 #import "HTMLNames.h"
 #import "LegacyNSPasteboardTypes.h"
@@ -49,6 +48,7 @@
 #import "LocalFrame.h"
 #import "LocalFrameView.h"
 #import "MutableStyleProperties.h"
+#import "NodeHTMLConverter.h"
 #import "PagePasteboardContext.h"
 #import "Pasteboard.h"
 #import "PasteboardStrategy.h"
