+++ Source/WebCore/page/writing-tools/WritingToolsController.mm
@@ -33,11 +33,11 @@
 #import "CompositeEditCommand.h"
 #import "DocumentInlines.h"
 #import "DocumentMarkerController.h"
+#import "EditingHTMLConverter.h"
 #import "Editor.h"
 #import "FocusController.h"
 #import "FrameSelection.h"
 #import "GeometryUtilities.h"
-#import "HTMLConverter.h"
 #import "IntelligenceTextEffectsSupport.h"
 #import "Logging.h"
 #import "NodeRenderStyle.h"
