+++ Source/WebKitLegacy/mac/WebCoreSupport/WebSelectionServiceController.mm
@@ -29,7 +29,7 @@
 
 #import "WebViewInternal.h"
 #import <WebCore/FrameSelection.h>
-#import <WebCore/HTMLConverter.h>
+#import <WebCore/NodeHTMLConverter.h>
 #import <WebCore/Page.h>
 #import <WebCore/Range.h>
 #import <WebCore/markup.h>
